+++ RELEASE_NOTES.md
@@ -2,7 +2,10 @@ Release Notes
 =============
 
 ## 1.6.19
+* Application Gateways: support for creating application gateways.
+* Container Service (AKS): support for various addons, including the application gateway ingress controller.
 * ExpressRoute: create authorization keys on newly created circuits.
+* Key Vaults: Add keys to new or existing key vaults.
 * ServiceBus: Allow Service Bus Queues/Topics/Subscriptions to be linked to unmanaged namespaces
 * ServiceBus: Allow adding custom dependencies to Subscriptions
 * WebApp/Functions: Adds 'ftp_state' for controlling FTP access for deployments.